Warning fix.

+++ fastlib/trunk/contrib/dongryel/thesis_research/core/parallel/distributed_dualtree_dfs_dev.h
@@ -119,7 +119,7 @@ void DistributedDualtreeDfs<DistributedProblemType>::AllToAllIReduce_(
         if((world_->size() == 1 &&
             distributed_tasks.num_remaining_tasks() < omp_get_num_threads()) ||
             (world_->size() > 1 &&
-             hashed_essential_reference_subtrees_to_send.size() <
+             static_cast<int>(hashed_essential_reference_subtrees_to_send.size()) <
              omp_get_num_threads())) {
           distributed_tasks.WalkReferenceTree(
             metric, problem_->global(), *world_, 2 * omp_get_num_threads(),
